% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{beamer}
\alias{beamer}
\title{Convert Rmd into a beamer presentation}
\usage{
beamer(pandoc_args = c("--listings"), slide_level = 2, ...)
}
\arguments{
\item{pandoc_args}{Pandoc arguments to be passed to \code{\link{pdf_output}};
defaults to "--listings"}

\item{slide_level}{Heading level defining individual slides; defaults to 2}

\item{...}{Arguments to be passed to \code{\link{pdf_output}}}
}
\description{
Convert Rmd into a beamer presentation
}
