% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_title.R
\name{cat_header_to_yaml}
\alias{cat_header_to_yaml}
\title{Concatenate and print header info as YAML parameter}
\usage{
cat_header_to_yaml(
  title,
  subtitle = "FALSE",
  other = "FALSE",
  id,
  date = format(Sys.Date(), "\%B \%d, \%Y"),
  author = "Omkar A. Katta"
)
}
\arguments{
\item{title}{Title of document}

\item{subtitle}{Subtitle of document}

\item{other}{Other header information}

\item{id}{Unique label/identifier of the document; useful for multi-document
formats}

\item{date}{Date of document; defaults to date of last render}

\item{author}{Author of document; defaults to "Omkar A. Katta"}
}
\value{
YAML header with the pertinent information to be used
by pandoc template
}
\description{
Single-document formats use this function to record the header as YAML
paramaters.
}
\details{
Multi-document formats require the title of the Rmd file as a heading,
not as a title. But single-document formats require the title of the Rmd
to be the title of the document. Hence, this function will \code{cat} the
header information as YAML parameters.
}
