% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Initialize {oakdown} template in current directory}
\usage{
init(
  template = "note",
  name = template,
  ext = ".rmd",
  use_gitignore = TRUE,
  overwrite_gitignore = FALSE
)
}
\arguments{
\item{template}{Name of {oakdown} template; defaults to "note"}

\item{name}{Name of file; defaults to \code{template}}

\item{ext}{Extension of file; defaults to ".rmd"}

\item{use_gitignore}{If TRUE (default), create .gitignore if one does not
already exist.}

\item{overwrite_gitignore}{If \code{use_gitignore} is TRUE and .gitignore
exists, overwrite the current .gitignore with the {oakdown} .gitignore}
}
\description{
Create files from {oakdown} template
}
\details{
If \code{use_gitignore} is TRUE, this function creates a .gitignore
if one does not already exist in the current directory.
This function is a wrapper for \code{rmarkdown::draft()}.
}
