% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_jot}
\alias{init_jot}
\title{Initialize jot template}
\usage{
init_jot(name, ext = ".rmd", use_gitignore = TRUE, overwrite_gitignore = FALSE)
}
\arguments{
\item{name}{Name of file}

\item{ext}{Extension of file; defaults to ".rmd"}

\item{use_gitignore}{If TRUE (default), create .gitignore if one does not
already exist.}

\item{overwrite_gitignore}{If \code{use_gitignore} is TRUE and .gitignore
exists, overwrite the current .gitignore with the {oakdown} .gitignore}
}
\description{
In the current directory, create an Rmd file called \code{name} based
on the skeleton file in {oakdown}'s note template.
If \code{use_gitignore} is TRUE and no .gitignore exists in the current
directory, a .gitignore file will be created.
If a .gitignore does exist, no .gitignore will be created unless
\code{overwrite_gitignore} is TRUE.
}
\seealso{
\code{\link{init}}
}
