% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_note}
\alias{init_note}
\title{Initialize note template}
\usage{
init_note(name, use_gitignore = TRUE, overwrite_gitignore = FALSE)
}
\arguments{
\item{name}{Name of Rmd file; defaults to \code{template}}

\item{use_gitignore}{If TRUE (default), create .gitignore if one does not
already exist.}

\item{overwrite_gitignore}{If \code{use_gitignore} is TRUE and .gitignore
exists, overwrite the current .gitignore with the {oakdown} .gitignore}
}
\description{
Initialize note template
}
\seealso{
\code{\link{init}}
}
