% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{pdf_quicknote}
\alias{pdf_quicknote}
\title{Convert Rmd into a quicknote}
\usage{
pdf_quicknote(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{pdf_output}}}
}
\description{
Convert Rmd into a quicknote
}
