% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_title.R
\name{print_title}
\alias{print_title}
\title{Print header of document}
\usage{
print_title(
  title,
  subtitle = "FALSE",
  other = "FALSE",
  date = format(Sys.Date(), "\%B \%d, \%Y"),
  author = "Omkar A. Katta",
  id
)
}
\arguments{
\item{title}{Title of document}

\item{subtitle}{Subtitle of document}

\item{other}{Other header information}

\item{date}{Date of document; defaults to date of last render}

\item{author}{Author of document; defaults to "Omkar A. Katta"}

\item{id}{Unique label/identifier of the document; useful for multi-document
formats}
}
\value{
Either a yaml header with the pertinent information to be used
by pandoc template or a series of TeX commands
}
\description{
Print the header conditional on the format
}
\details{
Different formats require different ways to print the header of the document.
Right now, only the single-document note format is supported.
}
