% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_fcns.R
\name{render_beamer}
\alias{render_beamer}
\title{Render Rmd to a Beamer presentation}
\usage{
render_beamer(
  input = "beamer",
  ext = "rmd",
  output = paste0("rendered_", input),
  output_dir = "_render",
  pre_script = resources("all", "settings.r"),
  clean = TRUE,
  ...
)
}
\arguments{
\item{input}{Input Rmd, to be passed to
\code{bookdown::render_book()}}

\item{ext}{Extension of input; defaults to "rmd"}

\item{output}{Name of the output PDF; defaults to "rendered_\code{input}"}

\item{output_dir}{Name of output directory}

\item{pre_script}{Name of R script to run before rendering}

\item{clean}{Argument to be passed to \code{tinytex.clean} option;
if TRUE (default), delete intermediaries, e.g., *.aux, *.toc, etc.}

\item{...}{Arguments to be passed to \code{rmarkdown::render}}
}
\value{
A PDF beamer presentation
}
\description{
Render Rmd to a Beamer presentation
}
