% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_fcns.R
\name{render_jot}
\alias{render_jot}
\title{Render Rmd to a PDF jot}
\usage{
render_jot(
  input = "jot",
  ext = "rmd",
  output = paste0("rendered_", input),
  output_dir = "_render",
  output_format = oakdown::pdf_jot(),
  pre_script = resources("all", "settings.r"),
  config_file = "jot-bookdown.yml",
  save_config_file = FALSE,
  clean = TRUE,
  clean_bookdown = "",
  title_fcn = cat_header_to_yaml,
  ...
)
}
\arguments{
\item{input}{Input file, to be passed to
\code{bookdown::render_book()}}

\item{ext}{Extension of input; defaults to "rmd"}

\item{output}{Name of the output PDF; defaults to "rendered_\code{input}"}

\item{output_dir}{Name of output directory}

\item{output_format}{Name of output format; defaults to
\code{oakdown::pdf_jot}}

\item{pre_script}{Name of R script to run before rendering}

\item{config_file}{Book configuration file; if file does not exist, then
create it}

\item{save_config_file}{Save generated book configuration file}

\item{clean}{Argument to be passed to \code{tinytex.clean} option;
if TRUE (default), delete intermediaries, e.g., *.aux, *.toc, etc.}

\item{clean_bookdown}{Argument to be passed to bookdown configuration;
vector of files to be removed}

\item{title_fcn}{Function to interpret header information in Rmd}

\item{...}{Arguments to be passed to \code{bookdown::render_book}}
}
\value{
A PDF jot with the name and directory as specified by
\code{config_file}
}
\description{
This function sets options, generates a bookdown configuration file,
and wraps \code{bookdown::pdf_book}.
Note that in the bookdown configuration process, this function, by default,
specifies the R script to run before the "chapter", which in our case is
is the jot.
}
