% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{template_resources}
\alias{template_resources}
\title{Find resources within {oakdown} template}
\usage{
template_resources(template, file = "pandoc.tex", subdir = "resources")
}
\arguments{
\item{template}{Name of template}

\item{file}{Name of file in \code{subdir} directory of \code{template}}

\item{subdir}{Name of subdirectory in \code{template}}
}
\value{
File path of \code{file} located in {oakdown}'s
\code{template} directory
}
\description{
Find resources within {oakdown} template
}
