% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_fcns.R
\name{render_note}
\alias{render_note}
\title{Render Rmd to a PDF note}
\usage{
render_note(
  input = ".",
  config_file = "note-bookdown.yml",
  clean = FALSE,
  title_fcn = cat_header_to_yaml,
  ...
)
}
\arguments{
\item{input}{Input filename or directory, to be passed to
\code{bookdown::render_book()}}

\item{config_file}{Book configuration file}

\item{clean}{Argument to be passed to \code{tinytex.clean} option;
if FALSE (default), keep intermediary files}

\item{title_fcn}{Function to interpret header information in Rmd}

\item{...}{Arguments to be passed to \code{bookdown::render_book}}
}
\value{
A PDF note with the name and directory as specified by
\code{config_file}
}
\description{
Render Rmd to a PDF note
}
